% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProcessSpace.R
\name{addProcessSpace}
\alias{addProcessSpace}
\title{addProcessSpace}
\usage{
addProcessSpace(transectObject)
}
\arguments{
\item{transectObject}{Object created by the \code{generateCrossSections} function}
}
\value{
transectObject that can be passed on to other functions or passed to exportSpatials
}
\description{
addProcessSpace
}
