% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStreamChannels.R
\name{addStreamChannels}
\alias{addStreamChannels}
\title{addStreamChannels}
\usage{
addStreamChannels(
  transectObject,
  sampleDensity = units::as_units(3, "m"),
  rasterDir = "GeoData/Raster/ChildsDEM_m.tif",
  streamChannelFile = "GeoData/STREAM CHANNELS.shp"
)
}
\arguments{
\item{rasterDir}{}
}
\value{

}
\description{
addStreamChannels
}
