% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTopoLines.R
\name{addTopoLines}
\alias{addTopoLines}
\title{addTopoLines}
\usage{
addTopoLines(transectObject, rasterDir = "GeoData/Raster/ChildsDEM_m.tif", ...)
}
\arguments{
\item{rasterDir}{}
}
\value{

}
\description{
addTopoLines
}
