% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{kml_compress_fixed}
\alias{kml_compress_fixed}
\title{kml_compress_fixed This function is based on the kml_compress from plotKML
but fixes a bug with including files}
\usage{
kml_compress_fixed(file.name, files = "", removeTemps = TRUE)
}
\arguments{
\item{removeTemps}{}
}
\value{

}
\description{
kml_compress_fixed This function is based on the kml_compress from plotKML
but fixes a bug with including files
}
