% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalElevation.R
\name{longitudinalElevation}
\alias{longitudinalElevation}
\title{longitudinalElevation}
\usage{
longitudinalElevation(
  transectObject,
  rasterDir = "GeoData/Raster/ChildsDEM_m.tif",
  makePlot = TRUE,
  returnData = TRUE,
  plotFileName = "longPlot.png"
)
}
\arguments{
\item{plotFileName}{}
}
\value{

}
\description{
longitudinalElevation
}
