% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPlotter.R
\name{rasterPlotter}
\alias{rasterPlotter}
\title{rasterPlotter}
\usage{
rasterPlotter(
  transectObject,
  rasterDir = "GeoData/Raster/ChildsDEM_m.tif",
  interpSmoothness = "Smooth"
)
}
\arguments{
\item{interpSmoothness}{}
}
\value{

}
\description{
rasterPlotter
}
