% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStreamChannels.R
\name{addStreamChannels}
\alias{addStreamChannels}
\title{addStreamChannels}
\usage{
addStreamChannels(
  transectObject,
  sampleDensity = units::as_units(6, "m"),
  rasterDir = "GeoData/Raster/ChildsDEM_m.tif",
  streamDir = "GeoData/STREAM CHANNELS.shp",
  ...
)
}
\arguments{
\item{transectObject}{Object created by the \code{generateCrossSections} function}

\item{sampleDensity}{The precision of the}

\item{rasterDir}{The file location of the DEM file.}
}
\value{

}
\description{
addStreamChannels
}
