% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allAtOnce.R
\name{allAtOnce}
\alias{allAtOnce}
\title{allAtOnce}
\usage{
allAtOnce(
  transectObject,
  outputFilename,
  detrendElevs = TRUE,
  returnObject = FALSE,
  doExportSpatial = FALSE
)
}
\arguments{
\item{doExportSpatial}{}
}
\value{

}
\description{
allAtOnce
}
