% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildXSectionPlot.R
\name{buildXSectionPlot}
\alias{buildXSectionPlot}
\title{buildXSectionPlot}
\usage{
buildXSectionPlot(
  transectObject,
  plotFileName = "test.png",
  returnData = TRUE,
  verticalCutoff = 4
)
}
\arguments{
\item{verticalCutoff}{}
}
\value{

}
\description{
buildXSectionPlot
}
