% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpatials.R
\name{exportSpatials}
\alias{exportSpatials}
\title{exportSpatials}
\usage{
exportSpatials(
  transectObject,
  sectionName = "temp",
  addXSectionPlots = TRUE,
  rasterPaletteLims = c(-2, 2)
)
}
\arguments{
\item{rasterPaletteLimits}{}
}
\value{

}
\description{
exportSpatials
}
