% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCrossSections.R
\name{generateCrossSections}
\alias{generateCrossSections}
\title{Generate Stream Cross Sections}
\usage{
generateCrossSections(
  streamChannel,
  getSatImage = TRUE,
  googleZoom = 14,
  cut1Dir = "W",
  cut2Dir = "E",
  xSectionLength = 100,
  xSectionDensity = units::as_units(100, "m")
)
}
\arguments{
\item{xSectionDensity}{}
}
\value{

}
\description{
Generate Stream Cross Sections
}
\examples{

}
